// front_abcount-main/src/app/dto/enterprise.dto.ts
export interface EnterpriseDto {
    address: string;
    ciRepresentative: string;
    companyName: string;
    emailRepresentative: string;
    legalRepresentative: string;
    logoUuid: string;
    nit: string;
    numberEmployee: string;
    numberRegistration: string;
    numberRepresentative: string;
    rubro: string;
}